# nodejs-server-server-generated/api/openapi.yaml
openapi: 3.0.1
info:
  title: Online Order API
  description: A basic API for working with the Swagger tools.
  version: 1.0.0
servers:
- url: /
paths:
  /orders:
    summary: Get all of the orders data.
    description: This path is used to retrieve all of the orders data from the orders.json
      file.
    get:
      tags:
      - Orders
      summary: Gets the order data
      description: Retrieve the order information from the orders.json file.
      operationId: get_orders
      responses:
        "200":
          description: Success
          content:
            application/json:
              examples:
                orders:
                  value: "{\"orders\":[{\"name\":\"Carey Maynard\",\"id\":\"001\"\
                    ,\"state\":\"pending\"},{\"name\":\"Angelo Ayala\",\"id\":\"002\"\
                    ,\"state\":\"canceled\"},{\"name\":\"Regina Yates\",\"id\":\"\
                    003\",\"state\":\"pending\"},{\"name\":\"Elliott Mcclure\",\"\
                    id\":\"004\",\"state\":\"pending\"}]}"
        default:
          description: Default error sample response
      x-swagger-router-controller: Orders
  /neworder:
    summary: Add new order.
    description: This path is used to add a new order to the orders.json file.
    post:
      tags:
      - Post Order
      summary: Add new order data
      description: Add a new order to the orders.json file.
      operationId: ""
      requestBody:
        description: A new order object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        default:
          description: Default error sample response
      x-swagger-router-controller: PostOrder
  /update/{id}:
    summary: Update the status of an order.
    description: This path is used to change the status of an order matching the provided
      id. It modifies the state attribute of each order.
    put:
      tags:
      - Update Order
      summary: Update the state of an order
      description: Update the state of an order with a matching id in the orders.json
        file.
      operationId: ""
      parameters:
      - name: id
        in: path
        description: The id of the order.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        default:
          description: Default error sample response
      x-swagger-router-controller: UpdateOrder
  /delete/{id}:
    summary: Delete an order.
    description: This path is used to delete an order with a matching id.
    delete:
      tags:
      - Delete Order
      summary: Deletes an order
      description: Delete an order with a matching id from the orders.json file.
      operationId: ""
      parameters:
      - name: id
        in: path
        description: The id of the order.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        default:
          description: Default error sample response
      x-swagger-router-controller: DeleteOrder
components:
  schemas:
    Order:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        state:
          type: string
      xml:
        name: Order